refactor: move to rtc folder
