+++ src/CMakeLists.txt
@@ -7,13 +7,10 @@ add_subdirectory(codecs/h264)
 add_subdirectory(codecs/v4l2)
 add_subdirectory(signaling)
 add_subdirectory(recorder)
+add_subdirectory(rtc)
 
 add_library(${PROJECT_NAME}
-    conductor.cpp
-    customized_video_encoder_factory.cpp
-    data_channel_subject.cpp
     parser.cpp
-    rtc_peer.cpp
 )
 
-target_link_libraries(${PROJECT_NAME} PUBLIC track capturer v4l2_codecs signaling recorder common)
+target_link_libraries(${PROJECT_NAME} PUBLIC rtc signaling recorder common)
