+++ src/rtc/customized_video_encoder_factory.cpp
@@ -1,4 +1,4 @@
-#include "customized_video_encoder_factory.h"
+#include "rtc/customized_video_encoder_factory.h"
 #include "codecs/v4l2/v4l2_h264_encoder.h"
 
 #include <modules/video_coding/codecs/av1/av1_svc_config.h>
